fix GetVectorDistanceWithLimitedWeight

+++ vvcx/Order.cs
@@ -10,18 +10,19 @@ namespace vvcx
 
     class Order
     {
-        public Order(Shop orderShop, List<Tuple<Product, double>> orderedProtuctList)
+        // int - ilość danego produktu do zamówienia
+        public Order(Shop orderShop, List<Tuple<Product, int>> orderedProtuctList)
         {
             shop = orderShop;
             orderedProducts = orderedProtuctList;
         }
 
         public string Name { get { return shop.Name; } }
-        public List<Tuple<Product, double>> OrderedProducts { get { return orderedProducts; } }
+        public List<Tuple<Product, int>> OrderedProducts { get { return orderedProducts; } }
         public Shop Shop { get { return shop; } }
 
         private Shop shop;
-        private List<Tuple<Product, double>> orderedProducts;
+        private List<Tuple<Product, int>> orderedProducts;
 
     }
 }
